requirements.txt
APScheduler==3.6.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.13
django-apscheduler==0.2.13
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-redis-cache==2.0.0
django-toolbelt==0.0.1
djangorestframework==3.9.2
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.3
parsel==1.5.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pytz==2018.9
queuelib==1.5.0
redis==3.2.1
requests==2.21.0
rq==0.13.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
static3==0.7.0
Twisted==18.9.0
tzlocal==1.5.1
urllib3==1.24.1
w3lib==1.20.0
whitenoise==4.1.2
zope.interface==4.6.0
